cli/pyproject.toml
[tool.poetry]
name = "moveboxtracker"
description = "moving box tracker and label generator"
authors = [ "Ian Kluft <ko6yq.dev+python@gmail.com>" ]
license = "GPL-3.0-or-later"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX",
    "Intended Audience :: End Users/Desktop"
]
version = "0.0.0" # placeholder for poetry-dynamic-versioning

[tool.poetry.dependencies]
python = "^3.10"
colorlookup = "^0.2"
weasyprint = ">= 55, < 58"
qrcode = "^7.0"
pyxdg = ">=0.27, < 2.0"
SQLAlchemy = "^2.0"

[tool.poetry.dev-dependencies]
flake8 = ">= 3.0.0, <7.0"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
#build-backend = "poetry.core.masonry.api"
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
